Animator Artillery + Upgrades Maya
Added in Unity under EthanScene the animator for the Artillery enemy.
Should work that when the script is attached it will switched. Being
that it is an idle to shoot and back there is little blending so
although I put it on no blend, just next frame move to shoot, it should
be fine.

Also added in the Maya scenes the Upgrade models. These are almost all
done besides the tri-shot upgrade which I am working on still