Pickups Integrated
Made changes to the Multi-shot Upgrade and then added them to Unity
under the Prefabs folder. Didn't want to mess to much so I just added
the meshes with animation on it, and added into EthanScene each model
with it's animator controller added so it can be quickly made into new
Prefabs.  Bomb pickup is the only one that has scripts copied from the
bomb scripts already done on the bomb Prefab found.