Tank Boss Animation and Pickup Size Bug Fix
Created the animations for the Tank Boss. There is an individual
animation for the left and right side of the artillery, turret, and
missile attacks as well as a version where both of the sides fire at the
same time. Also animated the top rotating. Any of these animations can
be used, and I added animation masks so several animations can play at
once (ex. a turret firing and an artillery). Also if for some reason the
attacks are so fast that they need to start before the restart of the
alternate side then the masks allow the second one to start with no
issues. The masks were done as part of the animator, made a new prefab
under new enemies with the animated version of the boss using the same
values and scripts as on the old one. Also believe I fixed the sizing
issue for the pickup prefabs.