Turrets. Cannonball, and Tri Drone
Sorry forgot to push. Basically put into Unity all of this, enjoy!
Turrets have colliders and tags, not sure what else is needed right now
for scripts or if they have been done. Tri-Drone has the same as I am
not sure if we are using the exact Kamakaze code